modified extract_data.py for 1+ cpus. Gpu flags added. Removed warning and cleaned code.
